+++ CHANGELOG.md
@@ -14,9 +14,11 @@ All notable changes to this project will be documented in this file.
 
 - Add missing role to read S3Connection and S3Bucket objects ([#112]).
 - Update annotation due to update to rust version ([#114]).
+- Update RBAC properties for OpenShift compatibility ([#126]).
 
 [#112]: https://github.com/stackabletech/spark-k8s-operator/pull/112
 [#114]: https://github.com/stackabletech/spark-k8s-operator/pull/114
+[#126]: https://github.com/stackabletech/spark-k8s-operator/pull/126
 
 ## [0.4.0] - 2022-08-03