+++ CHANGELOG.md
@@ -6,9 +6,9 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
-- Bumped image to `3.3.0-stackable0.2.0` in tests and docs ([#XX])
+- Bumped image to `3.3.0-stackable0.2.0` in tests and docs ([#145])
 
-[#XX]: https://github.com/stackabletech/spark-k8s-operator/pull/XX
+[#145]: https://github.com/stackabletech/spark-k8s-operator/pull/145
 
 ## [0.5.0] - 2022-09-06