+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Changed
+
+- Bumped image to `3.3.0-stackable0.2.0` in tests and docs ([#145])
+
+[#145]: https://github.com/stackabletech/spark-k8s-operator/pull/145
+
 ## [0.5.0] - 2022-09-06
 
 ### Added