+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.6.0] - 2022-11-07
+
 ### Changed
 
 - Bumped image to `3.3.0-stackable0.2.0` in tests and docs ([#145])